updated existing form to only include single date slot
