+++ firebase/src/components/Dialogs/ErrorDialog/index.tsx
@@ -1,16 +1,16 @@
 import React, { useState } from 'react'
-import Button from '@material-ui/core/Button';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import { makeStyles } from '@material-ui/core';
+import Button from '@material-ui/core/Button'
+import Dialog from '@material-ui/core/Dialog'
+import DialogActions from '@material-ui/core/DialogActions'
+import DialogContent from '@material-ui/core/DialogContent'
+import DialogContentText from '@material-ui/core/DialogContentText'
+import DialogTitle from '@material-ui/core/DialogTitle'
+import { makeStyles } from '@material-ui/core'
 
-const useStyles = makeStyles(theme => ({
+const useStyles = makeStyles((theme) => ({
     dialogContent: {
-        whiteSpace: "pre-wrap"
-    }
+        whiteSpace: 'pre-wrap',
+    },
 }))
 
 /**
@@ -23,10 +23,8 @@ export interface ErrorDialogProps {
 
 const WithErrorDialog = <P extends ErrorDialogProps>(
     Component: React.ComponentType<P>
- ): React.FC<Omit<P, keyof ErrorDialogProps>> => {
-
+): React.FC<Omit<P, keyof ErrorDialogProps>> => {
     const ComponentWithErrorDialog = (props: Omit<P, keyof ErrorDialogProps>) => {
-
         const classes = useStyles()
 
         var [open, setOpen] = useState(false)
@@ -43,11 +41,8 @@ const WithErrorDialog = <P extends ErrorDialogProps>(
 
         return (
             <>
-                <Dialog
-                    open={open}
-                    onClose={handleClose}
-                >
-                    <DialogTitle >{"Oops, something went wrong"}</DialogTitle>
+                <Dialog open={open} onClose={handleClose}>
+                    <DialogTitle>{'Something went wrong'}</DialogTitle>
                     <DialogContent>
                         <DialogContentText className={classes.dialogContent}>{errorMessage}</DialogContentText>
                     </DialogContent>
@@ -57,7 +52,7 @@ const WithErrorDialog = <P extends ErrorDialogProps>(
                         </Button>
                     </DialogActions>
                 </Dialog>
-                <Component { ...props as P } displayError={displayError} />
+                <Component {...(props as P)} displayError={displayError} />
             </>
         )
     }
